refactor: Consolidate FhirpathEvaluator factory methods into FhirpathEvaluators utility class
