--- fhirpath/src/main/java/au/csiro/pathling/fhirpath/execution/SingleFhirpathEvaluator.java
@@ -1,104 +1,0 @@
-/*
- * Copyright 2023 Commonwealth Scientific and Industrial Research
- * Organisation (CSIRO) ABN 41 687 119 230.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package au.csiro.pathling.fhirpath.execution;
-
-import au.csiro.pathling.fhirpath.FhirPath;
-import au.csiro.pathling.fhirpath.collection.Collection;
-import au.csiro.pathling.fhirpath.function.registry.FunctionRegistry;
-import au.csiro.pathling.io.source.DataSource;
-import ca.uhn.fhir.context.FhirContext;
-import jakarta.annotation.Nonnull;
-import java.util.List;
-import java.util.Map;
-import java.util.function.Supplier;
-import lombok.Value;
-import lombok.experimental.UtilityClass;
-import org.hl7.fhir.r4.model.Enumerations.ResourceType;
-
-
-@UtilityClass
-public class SingleFhirpathEvaluator {
-
-  @Value(staticConstructor = "of")
-  public static class SingleFactory implements FhirpathEvaluator.Factory {
-
-    @Nonnull
-    FhirContext fhirContext;
-
-    @Nonnull
-    DataSource dataSource;
-
-    @Override
-    @Nonnull
-    public FhirpathEvaluator create(@Nonnull final ResourceType subjectResource,
-        @Nonnull final FunctionRegistry<?> functionRegistry,
-        @Nonnull final Map<String, Collection> variables) {
-      return SingleFhirpathEvaluator.of(
-          subjectResource,
-          fhirContext,
-          functionRegistry,
-          variables,
-          dataSource
-      );
-    }
-  }
-
-  @Value
-  public static class SingleProvider implements FhirpathEvaluator.Provider {
-
-    @Nonnull
-    FhirContext fhirContext;
-
-    @Nonnull
-    FunctionRegistry<?> functionRegistry;
-
-    @Nonnull
-    Map<String, Collection> variables;
-
-    @Nonnull
-    DataSource dataSource;
-
-    @Nonnull
-    @Override
-    public FhirpathEvaluator create(
-        @Nonnull final ResourceType subjectResource,
-        @Nonnull final Supplier<List<FhirPath>> contextPaths) {
-      return SingleFhirpathEvaluator.of(
-          subjectResource,
-          fhirContext,
-          functionRegistry,
-          variables,
-          dataSource
-      );
-    }
-  }
-
-  @Nonnull
-  public static FhirpathEvaluator of(@Nonnull final ResourceType subjectResource,
-      @Nonnull final FhirContext fhirContext,
-      @Nonnull final FunctionRegistry<?> functionRegistry,
-      @Nonnull final Map<String, Collection> variables,
-      @Nonnull final DataSource dataSource) {
-    return new StdFhirpathEvaluator(
-        new SingleResourceResolver(subjectResource, fhirContext, dataSource),
-        functionRegistry,
-        variables
-    );
-  }
-  
-}
