+++ fhirpath/src/test/java/au/csiro/pathling/fhirpathe/execution/FhirpathTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import au.csiro.pathling.encoders.FhirEncoders;
+import au.csiro.pathling.fhirpath.execution.FhirpathEvaluators.MultiEvaluatorProvider;
 import au.csiro.pathling.fhirpath.execution.FhirpathExecutor;
-import au.csiro.pathling.fhirpath.execution.MultiFhirpathEvaluator.ManyProvider;
 import au.csiro.pathling.fhirpath.function.registry.StaticFunctionRegistry;
 import au.csiro.pathling.fhirpath.parser.Parser;
 import au.csiro.pathling.io.source.DataSource;
@@ -70,16 +70,15 @@ class FhirpathTest {
       @Nonnull final ResourceType subjectResource,
       @Nonnull final String fhirExpression) {
 
-    final Dataset<Row> resultDataset = createEvaluator(subjectResource, dataSource)
+    return createEvaluator(subjectResource, dataSource)
         .evaluate(subjectResource, fhirExpression)
         .toIdValueDataset();
-    return resultDataset;
   }
 
   @Nonnull
   FhirpathExecutor createEvaluator(@Nonnull final ResourceType subjectResource,
       @Nonnull final DataSource datasource) {
-    return FhirpathExecutor.of(new Parser(), new ManyProvider(encoders.getContext(),
+    return FhirpathExecutor.of(new Parser(), new MultiEvaluatorProvider(encoders.getContext(),
         StaticFunctionRegistry.getInstance(), Map.of(), datasource));
   }
 
@@ -815,12 +814,10 @@ class FhirpathTest {
     final ObjectDataSource dataSource = getPatientsWithConditions();
 
     final UnsupportedOperationException ex = assertThrows(UnsupportedOperationException.class,
-        () -> {
-          evalExpression(dataSource,
-              ResourceType.CONDITION,
-              "Patient.reverseResolve(Condition.subject).count()"
-          );
-        });
+        () -> evalExpression(dataSource,
+            ResourceType.CONDITION,
+            "Patient.reverseResolve(Condition.subject).count()"
+        ));
     assertEquals("Not implemented - nested resolves for foreign resources", ex.getMessage());
   }
 }
